requirements.txt
#Realistic Vision
#torch==2.0.1
torch==2.6.0
torchvision==0.21.0
safetensors==0.3.1
diffusers==0.19.0
#transformers==4.30.2
transformers==4.34.0
accelerate==0.20.3
omegaconf==2.3.0
controlnet_aux==0.0.3
compel==2.0.3
tqdm==4.67.1
numpy==1.23.5

#Head Segmentation
gdown==5.2.0
hydra-core==1.3.2
loguru==0.7.3
matplotlib==3.10.0
pandas==2.2.3
pillow==11.1.0
pytorch-lightning==2.5.0.post0
seaborn==0.13.2
segmentation_models_pytorch==0.4.0
streamlit==1.41.1
wandb==0.19.5

#DeepFace
deepface==0.0.93
tf_keras==2.18.0
tensorflow==2.12.0
